+++ src/Probel.Lanceur.Core/Probel.Lanceur.Core.csproj
@@ -59,12 +59,12 @@
     <Compile Include="Entities\AliasName.cs" />
     <Compile Include="Entities\AliasSession.cs" />
     <Compile Include="ServicesImpl\CommandRunner.cs" />
-    <Compile Include="ServicesImpl\Handlers\_TextReplacementHandler.cs" />
-    <Compile Include="ServicesImpl\Handlers\_ArgumentHandler.cs" />
-    <Compile Include="ServicesImpl\Handlers\ClipboardHandler.cs" />
-    <Compile Include="ServicesImpl\Handlers\TextHandler.cs" />
-    <Compile Include="ServicesImpl\Handlers\UrlHandler.cs" />
-    <Compile Include="ServicesImpl\Handlers\WebHandler.cs" />
+    <Compile Include="ServicesImpl\ArgumentHandlers\_TextReplacementHandler.cs" />
+    <Compile Include="ServicesImpl\ArgumentHandlers\_ArgumentHandler.cs" />
+    <Compile Include="ServicesImpl\ArgumentHandlers\ClipboardHandler.cs" />
+    <Compile Include="ServicesImpl\ArgumentHandlers\TextHandler.cs" />
+    <Compile Include="ServicesImpl\ArgumentHandlers\UrlHandler.cs" />
+    <Compile Include="ServicesImpl\ArgumentHandlers\WebHandler.cs" />
     <Compile Include="ServicesImpl\ParameterResolver.cs" />
     <Compile Include="ServicesImpl\TraceLogger.cs" />
     <Compile Include="ServicesImpl\Wildcards.cs" />
