+++ src/Probel.Lanceur.Core/ServicesImpl/ParameterResolver.cs
@@ -1,6 +1,6 @@
 ﻿using Probel.Lanceur.Core.Entities;
 using Probel.Lanceur.Core.Services;
-using Probel.Lanceur.Core.ServicesImpl.Handlers;
+using Probel.Lanceur.Core.ServicesImpl.ArgumentHandlers;
 using System.Text.RegularExpressions;
 
 namespace Probel.Lanceur.Core.ServicesImpl
